requirements.txt
pymongo
sentence-transformers
python-dotenv
fastapi
uvicorn
pydantic
google-generativeai
openai
ddgs
python-multipart
pypdf
